Updating Event Items - adding in "new" filter for `get_list`.
When retrieving event items, the "viewed" status of an item can now be used as
a filter. If the `new` parameter is specified, results returned can be
restricted to either only *new* items (that is items that have not yet been
viewed, downloaded, etc) or only items that have been previously viewed. If the
`new` parameter is not passed, then no filter is applied and both new and viewed
items are returned.